chore(glob): Move Glob to common/glob.ts
