+++ src/common/module.ts
@@ -1,7 +1,8 @@
 /** @module common */ /** for typedoc */
 export * from "./common";
+export * from "./coreservices";
+export * from "./glob";
 export * from "./hof";
 export * from "./predicates";
-export * from "./coreservices";
 export * from "./queue";
 export * from "./trace";
